Skip file stress tests on Windows
git-svn-id: https://svn.apache.org/repos/asf/camel/trunk@1340831 13f79535-47bb-0310-9956-ffa450edef68

+++ camel-core/src/test/java/org/apache/camel/component/file/stress/FileAsyncStressTest.java
@@ -33,6 +33,11 @@ public class FileAsyncStressTest extends ContextTestSupport {
 
     @Override
     protected void setUp() throws Exception {
+        // do not test on windows
+        if (isPlatform("windows")) {
+            return;
+        }
+
         super.setUp();
         deleteDirectory("target/filestress");
         for (int i = 0; i < files; i++) {
@@ -41,6 +46,11 @@ public class FileAsyncStressTest extends ContextTestSupport {
     }
 
     public void testAsyncStress() throws Exception {
+        // do not test on windows
+        if (isPlatform("windows")) {
+            return;
+        }
+
         // start route when all the files have been written
         context.startRoute("foo");
